+++ src/index.d.ts
@@ -1,8 +1,0 @@
-interface hmacOptions {
-    algorithm: string;
-    identifier: string;
-    header: string;
-    maxInterval: number;
-    minInterval: number;
-}
-export default function hmac(secret: string, options?: hmacOptions): function;